Added force field camouflage & particle effects
Fixed force field upgrade values computation
Added force field liquid removal (wip)
Fixed force field not happening below relays

+++ src/main/java/cr0s/warpdrive/api/IForceFieldUpgrade.java
@@ -5,5 +5,5 @@ public interface IForceFieldUpgrade {
 	IForceFieldUpgradeEffector getUpgradeEffector();
 	
 	// Bonus provided to this upgrade category (can be positive or negative)
-	int getUpgradeValue();
+	float getUpgradeValue();
 }
